Simplify service
