+++ src/app/backend/resource/service/detail.go
@@ -17,49 +17,20 @@ package service
 import (
 	"log"
 
-	"github.com/kubernetes/dashboard/src/app/backend/api"
 	"github.com/kubernetes/dashboard/src/app/backend/errors"
-	metricapi "github.com/kubernetes/dashboard/src/app/backend/integration/metric/api"
-	"github.com/kubernetes/dashboard/src/app/backend/resource/common"
-	"github.com/kubernetes/dashboard/src/app/backend/resource/dataselect"
 	"github.com/kubernetes/dashboard/src/app/backend/resource/endpoint"
-	"github.com/kubernetes/dashboard/src/app/backend/resource/pod"
-	v1 "k8s.io/api/core/v1"
+	"k8s.io/api/core/v1"
 	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8sClient "k8s.io/client-go/kubernetes"
 )
 
 // Service is a representation of a service.
 type ServiceDetail struct {
-	ObjectMeta api.ObjectMeta `json:"objectMeta"`
-	TypeMeta   api.TypeMeta   `json:"typeMeta"`
-
-	// InternalEndpoint of all Kubernetes services that have the same label selector as connected Replication
-	// Controller. Endpoints is DNS name merged with ports.
-	InternalEndpoint common.Endpoint `json:"internalEndpoint"`
-
-	// ExternalEndpoints of all Kubernetes services that have the same label selector as connected Replication
-	// Controller. Endpoints is external IP address name merged with ports.
-	ExternalEndpoints []common.Endpoint `json:"externalEndpoints"`
+	// Extends list item structure.
+	Service `json:",inline"`
 
 	// List of Endpoint obj. that are endpoints of this Service.
 	EndpointList endpoint.EndpointList `json:"endpointList"`
-
-	// Label selector of the service.
-	Selector map[string]string `json:"selector"`
-
-	// Type determines how the service will be exposed.  Valid options: ClusterIP, NodePort, LoadBalancer
-	Type v1.ServiceType `json:"type"`
-
-	// ClusterIP is usually assigned by the master. Valid values are None, empty string (""), or
-	// a valid IP address. None can be specified for headless services when proxying is not required
-	ClusterIP string `json:"clusterIP"`
-
-	// List of events related to this Service
-	EventList common.EventList `json:"eventList"`
-
-	// PodList represents list of pods targeted by same label selector as this service.
-	PodList pod.PodList `json:"podList"`
 
 	// Show the value of the SessionAffinity of the Service.
 	SessionAffinity v1.ServiceAffinity `json:"sessionAffinity"`
@@ -69,9 +40,7 @@ type ServiceDetail struct {
 }
 
 // GetServiceDetail gets service details.
-func GetServiceDetail(client k8sClient.Interface, metricClient metricapi.MetricClient, namespace, name string,
-	dsQuery *dataselect.DataSelectQuery) (*ServiceDetail, error) {
-
+func GetServiceDetail(client k8sClient.Interface, namespace, name string) (*ServiceDetail, error) {
 	log.Printf("Getting details of %s service in %s namespace", name, namespace)
 	serviceData, err := client.CoreV1().Services(namespace).Get(name, metaV1.GetOptions{})
 	if err != nil {
@@ -84,18 +53,15 @@ func GetServiceDetail(client k8sClient.Interface, metricClient metricapi.MetricC
 		return nil, criticalError
 	}
 
-	podList, err := GetServicePods(client, metricClient, namespace, name, dsQuery)
-	nonCriticalErrors, criticalError = errors.AppendError(err, nonCriticalErrors)
-	if criticalError != nil {
-		return nil, criticalError
-	}
-
-	eventList, err := GetServiceEvents(client, dataselect.DefaultDataSelect, namespace, name)
-	nonCriticalErrors, criticalError = errors.AppendError(err, nonCriticalErrors)
-	if criticalError != nil {
-		return nil, criticalError
-	}
-
-	service := ToServiceDetail(serviceData, *eventList, *podList, *endpointList, nonCriticalErrors)
+	service := toServiceDetail(serviceData, *endpointList, nonCriticalErrors)
 	return &service, nil
+}
+
+func toServiceDetail(service *v1.Service, endpointList endpoint.EndpointList, nonCriticalErrors []error) ServiceDetail {
+	return ServiceDetail{
+		Service:         toService(service),
+		EndpointList:    endpointList,
+		SessionAffinity: service.Spec.SessionAffinity,
+		Errors:          nonCriticalErrors,
+	}
 }
