+++ src/app/backend/resource/service/list_test.go
@@ -15,6 +15,8 @@
 package service
 
 import (
+	"github.com/kubernetes/dashboard/src/app/backend/resource/endpoint"
+	"github.com/kubernetes/dashboard/src/app/backend/resource/pod"
 	"reflect"
 	"testing"
 
@@ -80,6 +82,87 @@ func TestGetServiceList(t *testing.T) {
 
 		if !reflect.DeepEqual(actual, c.expected) {
 			t.Errorf("GetServiceList(client) == got\n%#v, expected\n %#v", actual, c.expected)
+		}
+	}
+}
+
+func TestToServiceDetail(t *testing.T) {
+	cases := []struct {
+		service      *v1.Service
+		eventList    common.EventList
+		podList      pod.PodList
+		endpointList endpoint.EndpointList
+		expected     ServiceDetail
+	}{
+		{
+			service:      &v1.Service{},
+			eventList:    common.EventList{},
+			podList:      pod.PodList{},
+			endpointList: endpoint.EndpointList{},
+			expected: ServiceDetail{
+				TypeMeta:          api.TypeMeta{Kind: api.ResourceKindService},
+				ExternalEndpoints: []common.Endpoint{},
+			},
+		}, {
+			service: &v1.Service{
+				ObjectMeta: metaV1.ObjectMeta{
+					Name: "test-service", Namespace: "test-namespace",
+				}},
+			expected: ServiceDetail{
+				ObjectMeta: api.ObjectMeta{
+					Name:      "test-service",
+					Namespace: "test-namespace",
+				},
+				TypeMeta:          api.TypeMeta{Kind: api.ResourceKindService},
+				InternalEndpoint:  common.Endpoint{Host: "test-service.test-namespace"},
+				ExternalEndpoints: []common.Endpoint{},
+			},
+		},
+	}
+
+	for _, c := range cases {
+		actual := ToServiceDetail(c.service, c.eventList, c.podList, c.endpointList, nil)
+
+		if !reflect.DeepEqual(actual, c.expected) {
+			t.Errorf("ToServiceDetail(%#v) == \ngot %#v, \nexpected %#v", c.service, actual,
+				c.expected)
+		}
+	}
+}
+
+func TestToService(t *testing.T) {
+	cases := []struct {
+		service  *v1.Service
+		expected Service
+	}{
+		{
+			service: &v1.Service{}, expected: Service{
+				TypeMeta:          api.TypeMeta{Kind: api.ResourceKindService},
+				ExternalEndpoints: []common.Endpoint{},
+			},
+		}, {
+			service: &v1.Service{
+				ObjectMeta: metaV1.ObjectMeta{
+					Name: "test-service", Namespace: "test-namespace",
+				}},
+			expected: Service{
+				ObjectMeta: api.ObjectMeta{
+					Name:      "test-service",
+					Namespace: "test-namespace",
+				},
+				TypeMeta:          api.TypeMeta{Kind: api.ResourceKindService},
+				InternalEndpoint:  common.Endpoint{Host: "test-service.test-namespace"},
+				ExternalEndpoints: []common.Endpoint{},
+			},
+		},
+	}
+
+	for _, c := range cases {
+		actual := ToService(c.service)
+
+		if !reflect.DeepEqual(actual, c.expected) {
+			t.Errorf("ToService(%#v) == \ngot %#v, \nexpected %#v", c.service, actual,
+				c.expected)
 		}
 	}
 }
